# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: frpmgr-server API
  description: |
    Headless FRP client manager. Manages multiple frpc instances inside a
    single process and exposes the full management surface (CRUD,
    lifecycle, status, logs, import/export) through HTTP. Real-time
    events (state changes, proxy status, log lines) are pushed via
    WebSocket at /api/v1/events.
  version: 1.0.0
servers:
  - url: http://localhost:18080
security:
  - BearerAuth: []
tags:
  - name: system
  - name: configs
  - name: proxies
  - name: lifecycle
  - name: logs
  - name: import-export
  - name: misc
paths:
  /api/v1/health:
    get:
      tags: [system]
      security: []
      summary: Liveness probe
      responses:
        '200':
          description: OK
  /api/v1/ui/branding:
    get:
      tags: [ui]
      security: []
      summary: 读取 UI 品牌（品牌名 + 浏览器标题）
      description: >
        返回生效品牌（未自定义的字段回退默认值）。公开端点，便于登录页与
        浏览器 <title> 在登录前即可渲染自定义值。
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Branding' }
    put:
      tags: [ui]
      summary: 更新 UI 品牌（需鉴权）
      description: >
        两个字段均可选：省略的字段保留当前存储值，显式空串则重置为默认。
        值会被 trim 并限长（app_name ≤ 40、html_title ≤ 120 字符）。
        返回更新后的生效品牌。
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Branding' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Branding' }
  /api/v1/version:
    get:
      tags: [system]
      summary: Version information
      responses:
        '200': { description: OK }
  /api/v1/version/check:
    get:
      tags: [system]
      summary: Check the latest GitHub release vs the running version
      parameters:
        - in: query
          name: force
          required: false
          schema: { type: integer, enum: [1] }
          description: Bypass the ~1h server-side cache when set to 1.
      responses:
        '200':
          description: >
            Version comparison result. Always returns current/frp/
            deployment_mode/self_update_enabled/has_update/can_self_update/
            reason; latest/changelog/html_url/published_at are present only
            when the GitHub query succeeded; check_error only on failure.
  /api/v1/system/update:
    post:
      tags: [system]
      summary: Launch an in-place self-update and restart the daemon
      description: >
        Spawns a detached updater that downloads the latest release, swaps the
        binary and restarts the service. Returns 202 immediately; the client
        should poll /version until the version changes. Gated by
        FRPCMGR_SELF_UPDATE_ENABLED and only available for service deployments
        (not Docker/manual). Pass ?force=1 to reinstall the current latest.
      parameters:
        - in: query
          name: force
          required: false
          schema: { type: integer, enum: [1] }
          description: Reinstall even when already up to date.
      responses:
        '202': { description: Update started; service restarting }
        '403': { description: Self-update disabled by operator }
        '400': { description: Deployment does not support self-update }
        '409': { description: Already up to date (without force) }
        '502': { description: Could not query the latest release }
  /api/v1/system/update/log:
    get:
      tags: [system]
      summary: Read the current self-update log
      description: >
        Returns the tail of update.log so the web UI can stream the live update
        steps (download / extract / install / restart). Reset to a fresh header
        each time POST /system/update starts. body { content: string }.
      responses:
        '200': { description: 'OK — body { content: string }' }
  /api/v1/system/config:
    get:
      tags: [system]
      summary: Read runtime daemon settings (env defaults + meta overrides)
      description: >
        Returns { effective, env_default, overridden } for log_level,
        self_update_enabled, docs_enabled, cors_origins. effective = env default
        overlaid with the operator's meta.json overrides.
      responses:
        '200': { description: OK }
    put:
      tags: [system]
      summary: Update runtime daemon settings (applies live, no restart)
      description: >
        Each provided field sets/replaces that override; a field name in "reset"
        clears it back to the env default. log_level change re-applies to the
        live logger; docs/self-update/CORS are read per-request. Persisted to
        meta.json so it survives restarts.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                log_level: { type: string, enum: [trace, debug, info, warn, error] }
                self_update_enabled: { type: boolean }
                docs_enabled: { type: boolean }
                cors_origins: { type: array, items: { type: string } }
                reset: { type: array, items: { type: string, enum: [log_level, self_update_enabled, docs_enabled, cors_origins] } }
      responses:
        '200': { description: 'OK — same shape as GET' }
        '400': { description: Invalid log_level or empty cors_origins }

  /api/v1/backup/channels:
    get:
      tags: [backup]
      summary: List storage channels (secrets masked)
      description: >
        Returns { channels: [...] }. Each channel omits its secret and exposes a
        *_set boolean (secret_access_key_set / password_set) instead.
      responses: { '200': { description: OK } }
    post:
      tags: [backup]
      summary: Create a storage channel (s3 | webdav)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, kind]
              properties:
                name: { type: string }
                kind: { type: string, enum: [s3, webdav] }
                s3:
                  type: object
                  properties:
                    endpoint: { type: string, description: 'host[:port], no scheme' }
                    region: { type: string }
                    bucket: { type: string }
                    access_key_id: { type: string }
                    secret_access_key: { type: string }
                    prefix: { type: string }
                    use_ssl: { type: boolean }
                    path_style: { type: boolean }
                webdav:
                  type: object
                  properties:
                    base_url: { type: string }
                    username: { type: string }
                    password: { type: string }
                    prefix: { type: string }
      responses:
        '201': { description: Created (masked channel) }
        '400': { description: Validation error }
  /api/v1/backup/channels/test:
    post:
      tags: [backup]
      summary: Test an unsaved channel config (no write)
      description: >
        Body is a channel object. If it carries an id of an existing channel and
        leaves the secret blank, the stored secret is reused. Returns
        { ok, error } — a failed test is a 200 result, not an HTTP error.
      responses: { '200': { description: 'OK — { ok: boolean, error?: string }' } }
  /api/v1/backup/channels/{id}:
    put:
      tags: [backup]
      summary: Update a channel (blank secret = keep current)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: OK (masked channel) }
        '404': { description: Channel not found }
    delete:
      tags: [backup]
      summary: Delete a channel (409 if referenced by a schedule)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '200': { description: OK }
        '404': { description: Not found }
        '409': { description: Referenced by a schedule }
  /api/v1/backup/channels/{id}/test:
    post:
      tags: [backup]
      summary: Test a stored channel's connectivity
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: 'OK — { ok, error? }' } }
  /api/v1/backup/schedules:
    get:
      tags: [backup]
      summary: List backup schedules (with running flag + last_run)
      responses: { '200': { description: OK } }
    post:
      tags: [backup]
      summary: Create a backup schedule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, channel_id, cron]
              properties:
                name: { type: string }
                enabled: { type: boolean }
                cron: { type: string, description: '5-field cron or @descriptor' }
                channel_id: { type: string }
                path_template: { type: string }
                retention: { type: integer, description: 'keep newest N, 0 = unlimited' }
      responses:
        '201': { description: Created }
        '400': { description: Validation error / unknown channel }
  /api/v1/backup/schedules/{id}:
    put:
      tags: [backup]
      summary: Update a schedule
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: OK }, '400': { description: Invalid }, '404': { description: Not found } }
    delete:
      tags: [backup]
      summary: Delete a schedule
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: OK }, '404': { description: Not found } }
  /api/v1/backup/schedules/{id}/toggle:
    post:
      tags: [backup]
      summary: Enable / disable a schedule (reloads cron)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: OK }, '404': { description: Not found } }
  /api/v1/backup/schedules/{id}/run:
    post:
      tags: [backup]
      summary: Trigger a manual backup now (async)
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        '202': { description: Started }
        '404': { description: Not found }
        '409': { description: Already running }
  /api/v1/backup/runs:
    get:
      tags: [backup]
      summary: Backup run history (newest first)
      parameters: [{ name: limit, in: query, schema: { type: integer, default: 50 } }]
      responses: { '200': { description: 'OK — { runs: [...] }' } }

  /api/v1/configs:
    get:
      tags: [configs]
      summary: List configs
      responses: { '200': { description: OK } }
    post:
      tags: [configs]
      summary: Create config
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id, config]
              properties:
                id: { type: string }
                config: { $ref: '#/components/schemas/ClientConfigV1' }
      responses:
        '201': { description: Created }
        '409': { description: Already exists }
  /api/v1/configs/reorder:
    post:
      tags: [configs]
      summary: Persist display order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                order: { type: array, items: { type: string } }
      responses: { '204': { description: OK } }
  /api/v1/configs/{id}:
    get:
      tags: [configs]
      summary: Get config
      parameters: [ { in: path, name: id, required: true, schema: { type: string } } ]
      responses: { '200': { description: OK }, '404': { description: Not found } }
    put:
      tags: [configs]
      summary: Replace config
      parameters: [ { in: path, name: id, required: true, schema: { type: string } } ]
      responses: { '200': { description: OK } }
    patch:
      tags: [configs]
      summary: Merge-patch config (RFC 7396)
      parameters: [ { in: path, name: id, required: true, schema: { type: string } } ]
      responses: { '200': { description: OK } }
    delete:
      tags: [configs]
      summary: Delete config
      parameters: [ { in: path, name: id, required: true, schema: { type: string } } ]
      responses: { '204': { description: OK } }
  /api/v1/configs/{id}/duplicate:
    post:
      tags: [configs]
      summary: Duplicate a config under a new id
      parameters: [ { in: path, name: id, required: true, schema: { type: string } } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [new_id]
              properties: { new_id: { type: string } }
      responses: { '201': { description: Created } }
  /api/v1/configs/{id}/raw:
    get:
      tags: [configs]
      summary: Read raw TOML
      parameters: [ { in: path, name: id, required: true, schema: { type: string } } ]
      responses: { '200': { description: OK } }
    put:
      tags: [configs]
      summary: Write raw TOML/INI
      parameters: [ { in: path, name: id, required: true, schema: { type: string } } ]
      requestBody:
        required: true
        content:
          application/toml: { schema: { type: string } }
          text/plain:       { schema: { type: string } }
      responses: { '200': { description: OK } }
  /api/v1/configs/{id}/proxies:
    get:
      tags: [proxies]
      summary: List proxies with live status
      parameters: [ { in: path, name: id, required: true, schema: { type: string } } ]
      responses: { '200': { description: OK } }
    post:
      tags: [proxies]
      summary: Add proxy or visitor
      description: >
        Visitors are checked for a local-port conflict: any instance's visitor
        (including the same config) of the same protocol family (STCP/XTCP=TCP,
        SUDP=UDP) on the same bindAddr:bindPort yields 409 visitor_port_conflict.
        Also enforced on PUT /configs/{id} and PUT /configs/{id}/raw; import
        paths intentionally skip it (restore fidelity — conflicts surface at run).
      parameters: [ { in: path, name: id, required: true, schema: { type: string } } ]
      responses:
        '201': { description: Created }
        '409': { description: 'proxy_already_exists, or visitor_port_conflict (details: config_id/config_name/name/type/bind_addr/bind_port)' }
  /api/v1/configs/{id}/proxies/reorder:
    post:
      tags: [proxies]
      summary: Reorder proxies/visitors (persist display order)
      description: >
        Re-sort the config's combined proxy/visitor list to match the supplied
        full ordered list of names. Names absent from the list keep their current
        relative position at the end, so a stale order never drops a proxy.
      parameters: [ { in: path, name: id, required: true, schema: { type: string } } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [order]
              properties:
                order: { type: array, items: { type: string } }
      responses: { '204': { description: OK }, '404': { description: Config not found } }
  /api/v1/configs/{id}/proxies/batch-delete:
    post:
      tags: [proxies]
      summary: Delete multiple proxies/visitors in one request
      description: One Update (single hot-reload) instead of N individual deletes.
      parameters: [ { in: path, name: id, required: true, schema: { type: string } } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [names]
              properties:
                names: { type: array, items: { type: string } }
      responses:
        '200': { description: 'OK — body { deleted: integer }' }
        '400': { description: names is required }
        '404': { description: No matching proxies }
  /api/v1/configs/{id}/proxies/move:
    post:
      tags: [proxies]
      summary: Move proxies/visitors to another config (atomic)
      description: >
        Relocate the named proxies/visitors from this config to target_id. The
        destination is written first (adding), the source second (removing), so a
        mid-way failure can only leave duplicates — never lose a proxy.
      parameters: [ { in: path, name: id, required: true, schema: { type: string } } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [target_id, names]
              properties:
                target_id: { type: string }
                names: { type: array, items: { type: string } }
      responses:
        '200': { description: 'OK — body { moved: integer }' }
        '400': { description: Missing/invalid params (e.g. target equals source) }
        '404': { description: Source/target config not found, or no matching proxies }
        '409': { description: 'Name already exists in target — details.names lists the clashes' }
  /api/v1/configs/{id}/proxies/{name}:
    get:
      tags: [proxies]
      summary: Get proxy
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
        - { in: path, name: name, required: true, schema: { type: string } }
      responses: { '200': { description: OK } }
    put:
      tags: [proxies]
      summary: Update proxy
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
        - { in: path, name: name, required: true, schema: { type: string } }
      responses:
        '204': { description: OK }
        '404': { description: proxy_not_found }
        '409': { description: 'visitor_port_conflict — any instance''s same-family visitor (incl. same config) uses this bindAddr:bindPort' }
    delete:
      tags: [proxies]
      summary: Delete proxy
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
        - { in: path, name: name, required: true, schema: { type: string } }
      responses: { '204': { description: OK } }
  /api/v1/configs/{id}/proxies/{name}/toggle:
    post:
      tags: [proxies]
      summary: Toggle proxy enabled state
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
        - { in: path, name: name, required: true, schema: { type: string } }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties: { enabled: { type: boolean } }
      responses: { '204': { description: OK } }
  /api/v1/configs/{id}/start:
    post:
      tags: [lifecycle]
      parameters: [ { in: path, name: id, required: true, schema: { type: string } } ]
      responses: { '200': { description: OK } }
  /api/v1/configs/{id}/stop:
    post:
      tags: [lifecycle]
      parameters: [ { in: path, name: id, required: true, schema: { type: string } } ]
      responses: { '200': { description: OK } }
  /api/v1/configs/{id}/reload:
    post:
      tags: [lifecycle]
      parameters: [ { in: path, name: id, required: true, schema: { type: string } } ]
      responses: { '200': { description: OK } }
  /api/v1/configs/{id}/status:
    get:
      tags: [lifecycle]
      parameters: [ { in: path, name: id, required: true, schema: { type: string } } ]
      responses: { '200': { description: OK } }
  /api/v1/configs/{id}/logs:
    get:
      tags: [logs]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
        - { in: query, name: lines, schema: { type: integer, default: 200 } }
        - { in: query, name: offset, schema: { type: integer, default: 0 } }
      responses: { '200': { description: OK } }
    delete:
      tags: [logs]
      parameters: [ { in: path, name: id, required: true, schema: { type: string } } ]
      responses: { '204': { description: OK } }
  /api/v1/configs/{id}/logs/files:
    get:
      tags: [logs]
      parameters: [ { in: path, name: id, required: true, schema: { type: string } } ]
      responses: { '200': { description: OK } }
  /api/v1/configs/{id}/logs/tail:
    get:
      tags: [logs]
      summary: WebSocket — live log tail
      description: 'Upgrade to WS. Each frame is a JSON object with a "line" key.'
      parameters: [ { in: path, name: id, required: true, schema: { type: string } } ]
      responses: { '101': { description: Switching Protocols } }
  /api/v1/events:
    get:
      tags: [misc]
      summary: WebSocket — global event stream
      description: |
        Upgrade to WS. Each frame is a JSON Event. Send
        `{"action":"filter","types":[...],"config_ids":[...]}` to update
        subscription, or `{"action":"unfilter"}` to receive everything.
        Query params `types` and `config_ids` (csv) set the initial filter,
        and `since=<seq>` replays events from the ring buffer.
      responses: { '101': { description: Switching Protocols } }
  /api/v1/validate:
    post:
      tags: [misc]
      summary: Validate config without persisting
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/ClientConfigV1' } }
          application/toml: { schema: { type: string } }
      responses: { '200': { description: OK } }
  /api/v1/import/file:
    post:
      tags: [import-export]
      summary: Import single config (multipart upload)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: { type: string, format: binary }
                id:   { type: string }
      responses: { '200': { description: OK } }
  /api/v1/import/url:
    post:
      tags: [import-export]
      summary: Import from HTTP URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url: { type: string }
                id:  { type: string }
      responses: { '200': { description: OK } }
  /api/v1/import/text:
    post:
      tags: [import-export]
      summary: Import inline TOML/INI text
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id, text]
              properties:
                id:   { type: string }
                text: { type: string }
                format: { type: string, enum: [toml, ini] }
      responses: { '200': { description: OK } }
  /api/v1/import/zip:
    post:
      tags: [import-export]
      summary: Import a ZIP backup
      description: >
        Restores configs from the archive's *.toml/*.ini/*.conf. If the archive
        contains meta.json (produced by /export/all), the operator branding, the
        instance display order and the runtime system_config overrides (log level
        / self-update / docs / CORS) are re-applied too; response.branding_restored
        / order_restored / system_config_restored report whether each was.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: { type: string, format: binary }
      responses:
        '200': { description: 'OK — body { imported: string[], branding_restored: boolean, order_restored: boolean, system_config_restored: boolean }' }
  /api/v1/configs/{id}/export:
    get:
      tags: [import-export]
      summary: Download single config (TOML)
      parameters: [ { in: path, name: id, required: true, schema: { type: string } } ]
      responses: { '200': { description: OK } }
  /api/v1/export/all:
    get:
      tags: [import-export]
      summary: Download every config as a ZIP
      description: >
        ZIP with profiles/*.toml plus meta.json (branding + sort), so the backup
        is self-contained and /import/zip can restore branding on a new host.
      responses: { '200': { description: OK } }
  /api/v1/nathole/discover:
    post:
      tags: [misc]
      summary: NAT-hole / STUN discovery
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties: { stun_server: { type: string } }
      responses: { '200': { description: OK } }
  /api/v1/system/info:
    get:
      tags: [system]
      summary: Aggregate dashboard snapshot (host+cpu+mem+disk+net+conns+process)
      responses: { '200': { description: OK } }
  /api/v1/system/cpu:
    get:
      tags: [system]
      summary: CPU usage and topology
      parameters: [ { in: query, name: window, schema: { type: string, default: 200ms } } ]
      responses: { '200': { description: OK } }
  /api/v1/system/memory:
    get:
      tags: [system]
      summary: Virtual + swap memory
      responses: { '200': { description: OK } }
  /api/v1/system/disk:
    get:
      tags: [system]
      summary: Disk usage for root + data dir
      parameters: [ { in: query, name: paths, schema: { type: string } } ]
      responses: { '200': { description: OK } }
  /api/v1/system/network:
    get:
      tags: [system]
      summary: Per-interface cumulative byte/packet counters
      responses: { '200': { description: OK } }
  /api/v1/system/connections:
    get:
      tags: [system]
      summary: Host-wide socket counts + daemon-owned subset
      responses: { '200': { description: OK } }
  /api/v1/system/process:
    get:
      tags: [system]
      summary: Daemon process introspection (RSS, threads, goroutines, open files)
      responses: { '200': { description: OK } }
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
    Branding:
      description: >
        UI 品牌。app_name 用于侧边栏与登录页品牌名，app_subtitle 用于侧边栏与
        登录页副标题，html_title 用于浏览器标签 <title>。GET 返回生效值（空字段
        已回退默认）；PUT 时各字段可选（省略保留、空串重置为默认）。
      type: object
      properties:
        app_name:     { type: string, maxLength: 40,  example: "FRPC" }
        app_subtitle: { type: string, maxLength: 60,  example: "客户端管理面板" }
        html_title:   { type: string, maxLength: 120, example: "FRPC · 内网穿透客户端管理控制台" }
    ClientConfigV1:
      description: |
        Mirrors fatedier/frp v0.68 ClientCommonConfig plus an `frpmgr`
        extension block ({name, manualStart, autoDelete}). See the
        upstream FRP documentation for the full field list; the JSON
        keys here match the camelCase used by frp.
      type: object
      additionalProperties: true
    Event:
      type: object
      properties:
        seq:       { type: integer }
        type:      { type: string, enum: [instance.state, instance.error, proxy.status, config.changed, config.deleted, log.line] }
        config_id: { type: string }
        ts:        { type: string, format: date-time }
        data:      { type: object, additionalProperties: true }